last_frame_added
I added one last frame (I added the adding_last_frame function) because I thought that without it I would get a normal output and it would look like it was hand cut.

+++ polaroid_horizontal/to_polaroid_horizontal.py
@@ -8,15 +8,24 @@ from image_cut import *
 def apply_polaroid_effect(input_path, output_folder, polaroid_size, frame_sizes=(150, 100, 150, 250)):
     #right, top, left, bottom frame sizes
 
-    image = Image.open(input_path)#resmi açar
-    grayscale_image = image.convert("L")#resmi gri tonlamalı yapar
-    polaroid_image = ImageOps.expand(grayscale_image, border=frame_sizes, fill='white')#beyaz arka plana sahip polaroid resmi oluşturur
+    image = Image.open(input_path)
+    grayscale_image = image.convert("L")
+    polaroid_image = ImageOps.expand(grayscale_image, border=frame_sizes, fill='white')
     polaroid_image = polaroid_image.resize(polaroid_size)#tekrar boyutlandırır
     timestamp = datetime.now().strftime("%H%M%S")
 
     output_filename = os.path.join(output_folder, f"output_polaroid_{timestamp}.jpg")#resmin adını oluşturur
     polaroid_image.save(output_filename)#resmi kaydede
     print(f"Polaroid image saved to: {output_filename}")
+    return polaroid_image  # apply_polaroid_effect fonksiyonu artık polaroid resmini döndürüyor.
+
+def adding_last_frame(polaroid_image, output_folder, frame_sizes=(3, 3, 3, 3)):
+    # right, top, left, bottom frame sizes
+    polaroid_with_frame = ImageOps.expand(polaroid_image, border=frame_sizes, fill='black')
+    timestamp = datetime.now().strftime("%H%M%S")
+    output_filename = os.path.join(output_folder, f"output_polaroid_{timestamp}.jpg")
+    polaroid_with_frame.save(output_filename)
+    print(f"Last framed polaroid image saved to: {output_filename}")
 
 if __name__ == "__main__":
     input_image = "reverend mother 2.jpg"
@@ -32,9 +41,14 @@ if __name__ == "__main__":
 
     width, height = get_image_dimensions(input_path)
     image_dimensions = (width, height)
-    apply_polaroid_effect(input_path, output_folder, image_dimensions)
+    polaroid_image = apply_polaroid_effect(input_path, output_folder, image_dimensions)
+    adding_last_frame(polaroid_image, output_folder)
 
-    try:
-        os.remove(input_path)
-    except Exception as e:
-        print(f"Error: {e}")
+    if input_path is not None:
+        try:
+            os.remove(input_path)
+            print(f"{input_path} deleted.")
+        except Exception as e:
+            print(f"Error: {e}")
+    else:
+        print("Input path is None, file deletion skipped.")
